feat: reorganize io module, reorganize loaders and writers structure, everything moved to "old_" modules
