+++ src/imgtools/io/loaders/__init__.py
@@ -1,4 +1,4 @@
-from .loaders import (
+from .old_loaders import (
     BaseLoader,
     ImageCSVLoader,
     ImageFileLoader,
@@ -12,15 +12,6 @@ from .loaders import (
     read_dicom_seg,
     read_dicom_series,
     read_image,
-)
-from .writers import (
-    BaseSubjectWriter,
-    BaseWriter,
-    HDF5Writer,
-    ImageFileWriter,
-    MetadataWriter,
-    NumpyWriter,
-    SegNrrdWriter,
 )
 
 __all__ = [
@@ -37,11 +28,4 @@ __all__ = [
     "ImageTreeLoader",
     "ImageCSVLoader",
     "ImageFileLoader",
-    "BaseWriter",
-    "BaseSubjectWriter",
-    "ImageFileWriter",
-    "SegNrrdWriter",
-    "NumpyWriter",
-    "HDF5Writer",
-    "MetadataWriter",
 ]
