+++ src/imgtools/io/loaders/old_loaders.py
