+++ src/imgtools/io/writers/__init__.py
@@ -1,0 +1,19 @@
+from .old_writers import (
+    BaseSubjectWriter,
+    BaseWriter,
+    HDF5Writer,
+    ImageFileWriter,
+    MetadataWriter,
+    NumpyWriter,
+    SegNrrdWriter,
+)
+
+__all__ = [
+    "BaseWriter",
+    "BaseSubjectWriter",
+    "ImageFileWriter",
+    "SegNrrdWriter",
+    "NumpyWriter",
+    "HDF5Writer",
+    "MetadataWriter",
+]
